rename plugins files with suffix 'plugin'
