+++ packages/mondrian/src/plugin/brushworks/highlighter.plugin.ts
@@ -1,24 +1,22 @@
-import { BLEND_MODES } from "pixi.js";
 import {
   IMondrianData,
   IMondrianInteractData,
   IMondrianStateData,
   MondrianDataType,
-} from "../data-manager";
-import { BrushName } from "./brush-common";
-import { PencilBrushPlugin } from "./pencil-plugin";
-import { PluginType } from "./plugin";
+} from "../../data-manager";
+import { PluginType, BrushName } from "../base";
+import { PencilBrushPlugin } from "./pencil.plugin";
 
-export class EraserBrushPlugin extends PencilBrushPlugin {
+export class HighlighterBrushPlugin extends PencilBrushPlugin {
   static override Type = PluginType.ConsumerExcludesive;
 
-  static override PID = Symbol("eraser-plugin");
+  static override PID = Symbol("highlighter-plugin");
 
   static override predicate(data: IMondrianData | null): boolean {
     if (data === null) return false;
     if (data.type === MondrianDataType.SET_STATE) {
       if (data as IMondrianStateData) {
-        if (data.data.player.brush.brushName === BrushName.Eraser) {
+        if (data.data.player.brush.brushName === BrushName.Highlighter) {
           return true;
         }
       }
@@ -30,9 +28,9 @@ export class EraserBrushPlugin extends PencilBrushPlugin {
     if (!super.reactDragStart(data)) return false;
     this.handler.config({
       enableDiscrete: false,
-      canCacheAsBitmap: false,
+      canCacheAsBitmap: true,
     });
-    this.handler.g.blendMode = BLEND_MODES.ERASE;
+    this.handler.c.filters = [this.sharedAlphaFilter];
     return true;
   }
 }
