--- packages/mondrian/src/plugin/semi-circle-plugin.ts
@@ -1,43 +1,0 @@
-import {
-  IMondrianData,
-  IMondrianInteractData,
-  IMondrianStateData,
-  MondrianDataType,
-} from "../data-manager";
-import { BrushName } from "./brush-common";
-import { PluginType } from "./plugin";
-import { ShapePlugin } from "./shape-plugin";
-
-export class SemiCirclePlugin extends ShapePlugin {
-  static override Type = PluginType.ConsumerExcludesive;
-
-  static override PID = Symbol("semi-circle-plugin");
-
-  static override predicate(data: IMondrianData | null): boolean {
-    if (data === null) return false;
-    if (data.type === MondrianDataType.SET_STATE) {
-      if (data as IMondrianStateData) {
-        if (data.data.player.brush.brushName === BrushName.SemiCircle) {
-          return true;
-        }
-      }
-    }
-    return false;
-  }
-
-  override reactDragMove(data: IMondrianInteractData): boolean {
-    if (!super.reactDragMove(data)) return false;
-    this.handler.g.clear();
-    this.handler.lineStyle = { ...this.handler.lineStyle };
-    this.handler.g.arc(
-      this.shapeRect.ox + this.shapeRect.dx / 2,
-      this.shapeRect.oy + this.shapeRect.dy,
-      Math.abs(this.shapeRect.dx / 2),
-      Math.PI,
-      Math.PI * 2,
-      !(this.shapeRect.dy > 0)
-    );
-    this.handler.g.closePath();
-    return true;
-  }
-}
