+++ packages/mondrian/src/plugin/shapes/rectangle-plugin.ts
@@ -3,21 +3,21 @@ import {
   IMondrianInteractData,
   IMondrianStateData,
   MondrianDataType,
-} from "../data-manager";
-import { BrushName } from "./brush-common";
-import { PluginType } from "./plugin";
-import { ShapePlugin } from "./shape-plugin";
+} from "../../data-manager";
+import { BrushName } from "../base/brush-common";
+import { PluginType } from "../base/plugin";
+import { ShapePlugin } from "../base/shape.plugin";
 
-export class CirclePlugin extends ShapePlugin {
+export class RectanglePlugin extends ShapePlugin {
   static override Type = PluginType.ConsumerExcludesive;
 
-  static override PID = Symbol("circle-plugin");
+  static override PID = Symbol("rectangle-plugin");
 
   static override predicate(data: IMondrianData | null): boolean {
     if (data === null) return false;
     if (data.type === MondrianDataType.SET_STATE) {
       if (data as IMondrianStateData) {
-        if (data.data.player.brush.brushName === BrushName.Circle) {
+        if (data.data.player.brush.brushName === BrushName.Rectangle) {
           return true;
         }
       }
@@ -29,11 +29,11 @@ export class CirclePlugin extends ShapePlugin {
     if (!super.reactDragMove(data)) return false;
     this.handler.g.clear();
     this.handler.lineStyle = { ...this.handler.lineStyle };
-    this.getDrawShapeHandle(data).drawEllipse(
-      this.shapeRect.x + this.shapeRect.w / 2,
-      this.shapeRect.y + this.shapeRect.h / 2,
-      this.shapeRect.w / 2,
-      this.shapeRect.h / 2
+    this.getDrawShapeHandle(data).drawRect(
+      this.shapeRect.x,
+      this.shapeRect.y,
+      this.shapeRect.w,
+      this.shapeRect.h
     );
     return true;
   }
