+++ packages/mondrian/src/plugin/shapes/right-angle-trapezoid.plugin.ts
@@ -3,21 +3,23 @@ import {
   IMondrianInteractData,
   IMondrianStateData,
   MondrianDataType,
-} from "../data-manager";
-import { BrushName } from "./brush-common";
-import { PluginType } from "./plugin";
-import { ShapePlugin } from "./shape-plugin";
+} from "../../data-manager";
+import { BrushName } from "../base/brush-common";
+import { PluginType } from "../base/plugin";
+import { ShapePlugin } from "../base/shape.plugin";
 
-export class RightAngleTrianglePlugin extends ShapePlugin {
+export class RightAngleTrapezoidPlugin extends ShapePlugin {
   static override Type = PluginType.ConsumerExcludesive;
 
-  static override PID = Symbol("right-angle-triangle-plugin");
+  static override PID = Symbol("right-angle-trapezoid-plugin");
 
   static override predicate(data: IMondrianData | null): boolean {
     if (data === null) return false;
     if (data.type === MondrianDataType.SET_STATE) {
       if (data as IMondrianStateData) {
-        if (data.data.player.brush.brushName === BrushName.RightAngleTriangle) {
+        if (
+          data.data.player.brush.brushName === BrushName.RightAngleTrapezoid
+        ) {
           return true;
         }
       }
@@ -29,15 +31,31 @@ export class RightAngleTrianglePlugin extends ShapePlugin {
     if (!super.reactDragMove(data)) return false;
     this.handler.g.clear();
     this.handler.lineStyle = { ...this.handler.lineStyle };
+
+    const deltaHorizontal = Math.abs(
+      this.shapeRect.dy / Math.tan((60 / 180) * Math.PI)
+    );
+
+    // illegal shape
+    if (Math.abs(this.shapeRect.dx) < deltaHorizontal) {
+      // draw nothing
+      return false;
+    }
+
+    const directionX = this.shapeRect.dx / Math.abs(this.shapeRect.dx);
+
     this.getDrawShapeHandle(data).drawPolygon([
       // point 1
       this.shapeRect.ox,
       this.shapeRect.oy,
       // point 2
-      this.shapeRect.ox,
-      this.shapeRect.oy + this.shapeRect.dy,
+      this.shapeRect.ox + this.shapeRect.dx - directionX * deltaHorizontal,
+      this.shapeRect.oy,
       // point 3
       this.shapeRect.ox + this.shapeRect.dx,
+      this.shapeRect.oy + this.shapeRect.dy,
+      // point 4
+      this.shapeRect.ox,
       this.shapeRect.oy + this.shapeRect.dy,
       // point 1
       this.shapeRect.ox,
