+++ packages/mondrian/src/plugin/shapes/sphere.plugin.ts
@@ -5,21 +5,21 @@ import {
   IMondrianInteractData,
   IMondrianStateData,
   MondrianDataType,
-} from "../data-manager";
-import { BrushName } from "./brush-common";
-import { PluginType } from "./plugin";
-import { ShapePlugin } from "./shape-plugin";
+} from "../../data-manager";
+import { BrushName } from "../base/brush-common";
+import { PluginType } from "../base/plugin";
+import { ShapePlugin } from "../base/shape.plugin";
 
-export class CylinderPlugin extends ShapePlugin {
+export class SpherePlugin extends ShapePlugin {
   static override Type = PluginType.ConsumerExcludesive;
 
-  static override PID = Symbol("cylinder-plugin");
+  static override PID = Symbol("sphere-plugin");
 
   static override predicate(data: IMondrianData | null): boolean {
     if (data === null) return false;
     if (data.type === MondrianDataType.SET_STATE) {
       if (data as IMondrianStateData) {
-        if (data.data.player.brush.brushName === BrushName.Cylinder) {
+        if (data.data.player.brush.brushName === BrushName.Sphere) {
           return true;
         }
       }
@@ -50,61 +50,27 @@ export class CylinderPlugin extends ShapePlugin {
     });
     this.handler.lineStyle = { ...this.handler.lineStyle };
 
+    // set up points
+    const po = {
+      x: this.shapeRect.ox + this.shapeRect.dx / 2,
+      y: this.shapeRect.oy + this.shapeRect.dy / 2,
+    };
     const rx = this.shapeRect.w / 2;
-    const ry = this.shapeRect.w / 2 / 3;
+    const ry = this.shapeRect.h / 2;
 
-    if (Math.abs(this.shapeRect.dy) < ry * 2) {
-      return false;
-    }
+    this.handler.g.drawEllipse(po.x, po.y, rx, ry);
 
-    // set up points
-    const poOfNear = {
-      x:
-        this.shapeRect.ox +
-        (this.shapeRect.dx / Math.abs(this.shapeRect.dx)) * rx,
-      y:
-        this.shapeRect.oy +
-        (this.shapeRect.dy / Math.abs(this.shapeRect.dy)) * ry,
-    };
-
-    // set up points
-    const poOfFar = {
-      x:
-        this.shapeRect.ox +
-        (this.shapeRect.dx / Math.abs(this.shapeRect.dx)) * rx,
-      y:
-        this.shapeRect.oy +
-        this.shapeRect.dy -
-        (this.shapeRect.dy / Math.abs(this.shapeRect.dy)) * ry,
-    };
-
-    // draw straight line
-    this.handler.g
-      .moveTo(this.shapeRect.ox, poOfNear.y)
-      .lineTo(this.shapeRect.ox, poOfFar.y);
-    this.handler.g
-      .moveTo(this.shapeRect.ox + this.shapeRect.dx, poOfNear.y)
-      .lineTo(this.shapeRect.ox + this.shapeRect.dx, poOfFar.y);
+    const ryOfEllipse = ry / 3;
+    const ellipsePoints = 80;
 
     const needReverseY = this.shapeRect.dy < 0;
-
-    // draw visible ellipse
-    const poOfVisibleEllipse = poOfNear;
-    this.handler.g.drawEllipse(
-      poOfVisibleEllipse.x,
-      poOfVisibleEllipse.y,
-      rx,
-      ry
-    );
-
-    const ellipsePoints = 80;
 
     const interval = (Math.PI * 2) / ellipsePoints;
     const dashline = this.customizedDash;
     let isInUpperTail = true;
     for (let i = 0; i < Math.PI * 2; i += interval) {
-      const x0 = poOfFar.x - rx * Math.cos(i);
-      const y0 = poOfFar.y - ry * Math.sin(i);
+      const x0 = po.x - rx * Math.cos(i);
+      const y0 = po.y - ryOfEllipse * Math.sin(i);
       let g: Graphics | DashLine = this.handler.g;
       // toggle g object between dashline and normal grahpic based on radius and drag direction in Y-axis.
       if (i > Math.PI) {
