+++ packages/mondrian/src/plugin/shapes/stroke-plugin.ts
@@ -3,10 +3,10 @@ import {
   IMondrianStateData,
   MondrianDataType,
   IMondrianInteractData,
-} from "../data-manager";
-import { BrushName } from "./brush-common";
-import { PluginType } from "./plugin";
-import { ShapePlugin } from "./shape-plugin";
+} from "../../data-manager";
+import { BrushName } from "../base/brush-common";
+import { PluginType } from "../base/plugin";
+import { ShapePlugin } from "../base/shape.plugin";
 
 export class StrokePlugin extends ShapePlugin {
   static override Type = PluginType.ConsumerExcludesive;
